Marking test classes as integrated tests when they require ioc's
